・DynamicPathVariables実装中 ・コードが増えてきてめんどいので別ファイルに分けた
