+++ Assets/Mizore_Nekoyanagi/Util/DynamicPathVariable.cs
@@ -1,0 +1,13 @@
+﻿namespace MizoreNekoyanagi.PublishUtil.PackageExporter
+{
+    [System.Serializable]
+    public class DynamicPathVariable
+    {
+        public string key, value;
+
+        public DynamicPathVariable( string key, string value ) {
+            this.key = key;
+            this.value = value;
+        }
+    }
+}
