+++ Assets/Mizore_Nekoyanagi/Util/DynamicPathVariable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d88159e4e70b9b438099ef97d307259
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
