+++ Assets/Mizore_Nekoyanagi/Util/ExporterTexts.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e214029d0dc5688498113e394d795ce5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
