+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter.cs
@@ -7,65 +7,27 @@ using System.Text;
 using UnityEditor;
 #endif
 
-namespace MizoreNekoyanagi.PublishUtil
+namespace MizoreNekoyanagi.PublishUtil.PackageExporter
 {
     [CreateAssetMenu( menuName = "MizoreNekoyanagi/UnityPackageExporter" )]
-    public class MizoresPackageExporter : ScriptableObject
+    public class MizoresPackageExporter : ScriptableObject, ISerializationCallbackReceiver
     {
-        [System.Serializable]
-        public class PackagePrefsElement
-        {
-            [SerializeField]
-            private Object obj;
-            [SerializeField]
-            private string path;
-
-            public Object Object {
-                get {
-#if UNITY_EDITOR
-                    if ( obj == null && !string.IsNullOrEmpty( path ) ) {
-                        obj = AssetDatabase.LoadAssetAtPath<Object>( path );
-                    }
-#endif
-                    return obj;
-                }
-                set {
-#if UNITY_EDITOR
-                    //if ( obj != value ) {
-                    if ( value != null ) {
-                        path = AssetDatabase.GetAssetPath( value.GetInstanceID( ) );
-                    } else {
-                        path = string.Empty;
-                    }
-                    //}
-#endif
-                    obj = value;
-                }
-            }
-
-            public string Path {
-                get {
-#if UNITY_EDITOR
-                    if ( obj != null ) {
-                        path = AssetDatabase.GetAssetPath( obj.GetInstanceID( ) );
-                    }
-#endif
-                    return path;
-                }
-            }
-        }
-
         public List<PackagePrefsElement> objects = new List<PackagePrefsElement>( );
         public List<string> dynamicpath = new List<string>( );
+        [SerializeField]
+        DynamicPathVariable[] s_variables;
+        [System.NonSerialized]
+        public Dictionary<string, string> variables = new Dictionary<string, string>( );
         public PackagePrefsElement versionFile;
 
         public const string EXPORT_FOLDER_PATH = "MizorePackageExporter/";
         public const string EXPORT_LOG_NOT_FOUND = "[{0}] is not exists. The export has been cancelled.\n[{0}]は存在しません。エクスポートは中断されました。\n";
         public string ExportPath { get { return EXPORT_FOLDER_PATH + this.name + ExportVersion + ".unitypackage"; } }
-        public string ExportVersion { get { return versionFile == null || string.IsNullOrEmpty(versionFile.Path) ? "" : "-" + File.ReadAllText(versionFile.Path).Trim(); } }
+        public string ExportVersion { get { return versionFile == null || string.IsNullOrEmpty( versionFile.Path ) ? "" : "-" + File.ReadAllText( versionFile.Path ).Trim( ); } }
 
-        string ConvertDynamicPath( string path ) {
+        public string ConvertDynamicPath( string path ) {
             path = path.Replace( "%name%", name );
+            path = path.Replace( "%version%", ExportVersion );
             return path;
         }
         public void Export( ) {
@@ -112,454 +74,15 @@ namespace MizoreNekoyanagi.PublishUtil
 #endif
         }
 
-#if UNITY_EDITOR
-        [CustomEditor( typeof( MizoresPackageExporter ) ), CanEditMultipleObjects]
-        public class UnityPackageExporterEditor : Editor
-        {
-            public struct MinMax
-            {
-                public int min, max;
+        public void OnBeforeSerialize( ) {
+            s_variables = variables.Select( kvp => new DynamicPathVariable( kvp.Key, kvp.Value ) ).ToArray( );
+        }
 
-                public MinMax( int min, int max ) {
-                    this.min = min;
-                    this.max = max;
-                }
-
-                public static MinMax Create<T>( IEnumerable<T> list, System.Func<T, int> func ) {
-                    int min = list.Min( func );
-                    int max = list.Max( func );
-                    return new MinMax( min, max );
-                }
-            }
-
-            public const string TEXT_UNDO = "PackagePrefs";
-            public const string TEXT_OBJECTS = "Objects";
-            public const string TEXT_DYNAMIC_PATH = "Dynamic Path";
-            public const string TEXT_VERSION_FILE = "Version File";
-            public const string TEXT_BUTTON_EXPORT = "Export to unitypackage";
-            public const string TEXT_BUTTON_EXPORT_M = "Export to unitypackages";
-            public const string TEXT_BUTTON_OPEN = "Open";
-            public const string TEXT_DIFF_LABEL = "?";
-            public const string TEXT_DIFF_TOOLTIP = "Some values are different.\n一部のオブジェクトの値が異なっています。";
-            public const string TEXT_BUTTON_FOLDER = "Folder";
-            public const string TEXT_BUTTON_FILE = "File";
-
-            public string t_Undo => TEXT_UNDO;
-            public string t_Objects => TEXT_OBJECTS;
-            public string t_DynamicPath => TEXT_DYNAMIC_PATH;
-            public string t_VersionFile => TEXT_VERSION_FILE;
-            public string t_Button_ExportPackage => TEXT_BUTTON_EXPORT;
-            public string t_Button_ExportPackages => TEXT_BUTTON_EXPORT_M;
-            public string t_Button_Open => TEXT_BUTTON_OPEN;
-            public string t_Diff_Label => TEXT_DIFF_LABEL;
-            public string t_Diff_Tooltip => TEXT_DIFF_TOOLTIP;
-            public string t_Button_Folder => TEXT_BUTTON_FOLDER;
-            public string t_Button_File => TEXT_BUTTON_FILE;
-
-            public static string HelpBoxText;
-            public static MessageType HelpBoxMessageType;
-            Vector2 scroll;
-            MizoresPackageExporter t;
-            private void OnEnable( ) {
-                HelpBoxText = null;
-                t = target as MizoresPackageExporter;
-            }
-            static string ToAssetsPath( string path ) {
-                string datapath = Application.dataPath;
-                if ( path.StartsWith( datapath ) ) {
-                    path = path.Substring( datapath.Length - "Assets".Length );
-                }
-                return path;
-            }
-            public static void Resize<T>( List<T> list, int newSize, System.Func<T> newvalue ) {
-                if ( list.Count == newSize ) {
-                    return;
-                } else if ( list.Count < newSize ) {
-                    var array = new T[newSize - list.Count];
-                    if ( newvalue != null ) {
-                        for ( int i = 0; i < array.Length; i++ ) {
-                            array[i] = newvalue.Invoke( );
-                        }
-                    }
-                    list.AddRange( array );
-                } else {
-                    list.RemoveRange( newSize, list.Count - newSize );
-                }
-            }
-            public static void Resize<T>( List<T> list, int newSize ) {
-                if ( list.Count == newSize ) {
-                    return;
-                } else if ( list.Count < newSize ) {
-                    list.AddRange( new T[newSize - list.Count] );
-                } else {
-                    list.RemoveRange( newSize, list.Count - newSize );
-                }
-            }
-            string BrowseButtons( string text ) {
-                string result=text;
-                if ( GUILayout.Button( TEXT_BUTTON_FOLDER, GUILayout.Width( 50 ) ) ) {
-                    text = EditorUtility.OpenFolderPanel( null, t.ConvertDynamicPath( text ), null );
-                    text = ToAssetsPath( text );
-                    if ( string.IsNullOrEmpty( text ) == false ) {
-                        GUI.changed = true;
-                        result = text;
-                    }
-                }
-                if ( GUILayout.Button( TEXT_BUTTON_FILE, GUILayout.Width( 50 ) ) ) {
-                    text = EditorUtility.OpenFilePanel( null, t.ConvertDynamicPath( text ), null );
-                    text = ToAssetsPath( text );
-                    if ( string.IsNullOrEmpty( text ) == false ) {
-                        GUI.changed = true;
-                        result = text;
-                    }
-                }
-                return result;
-            }
-            /// <summary>
-            /// 複数オブジェクトの編集
-            /// </summary>
-            void EditMultiple( ) {
-                if ( targets.Length <= 1 ) return;
-                scroll = EditorGUILayout.BeginScrollView( scroll );
-                Undo.RecordObjects( targets, t_Undo );
-
-                var targetlist = targets.Select( v => v as MizoresPackageExporter );
-
-                // Targets
-                GUI.enabled = false;
-                foreach ( var item in targetlist ) {
-                    EditorGUILayout.ObjectField( item, typeof( MizoresPackageExporter ), false );
-                }
-                GUI.enabled = true;
-
-                // ↓ Objects
-                EditorGUILayout.Separator( );
-                EditorGUILayout.LabelField( t_Objects, EditorStyles.boldLabel );
-                MinMax objects_count = MinMax.Create( targetlist, v => v.objects.Count );
-                for ( int i = 0; i < objects_count.max; i++ ) {
-                    using ( var horizontalScope = new EditorGUILayout.HorizontalScope( ) ) {
-                        // 全てのオブジェクトの値が同じか
-                        bool same = i < objects_count.min && targetlist.All( v => t.objects[i].Object == v.objects[i].Object );
-
-                        if ( same ) {
-                            EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-                        } else {
-                            EditorGUILayout.LabelField( new GUIContent( t_Diff_Label, t_Diff_Tooltip ), GUILayout.Width( 30 ) );
-                        }
-
-                        EditorGUI.BeginChangeCheck( );
-                        Object obj;
-                        if ( same ) {
-                            obj = EditorGUILayout.ObjectField( t.objects[i].Object, typeof( Object ), false );
-                        } else {
-                            obj = EditorGUILayout.ObjectField( null, typeof( Object ), false );
-                        }
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            foreach ( var item in targetlist ) {
-                                Resize( item.objects, Mathf.Max( i + 1, item.objects.Count ), ( ) => new PackagePrefsElement( ) );
-                                item.objects[i].Object = obj;
-                                objects_count = MinMax.Create( targetlist, v => v.objects.Count );
-                                EditorUtility.SetDirty( item );
-                            }
-                        }
-
-                        EditorGUI.BeginChangeCheck( );
-                        string path;
-                        if ( same ) {
-                            path = EditorGUILayout.TextField( t.objects[i].Path );
-                        } else {
-                            path = EditorGUILayout.TextField( string.Empty );
-                        }
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            // パスが変更されたらオブジェクトを置き換える
-                            Object o = AssetDatabase.LoadAssetAtPath<Object>( path );
-                            if ( o != null ) {
-                                foreach ( var item in targetlist ) {
-                                    Resize( item.objects, Mathf.Max( i + 1, item.objects.Count ), ( ) => new PackagePrefsElement( ) );
-                                    item.objects[i].Object = o;
-                                    objects_count = MinMax.Create( targetlist, v => v.objects.Count );
-                                    EditorUtility.SetDirty( item );
-                                }
-                            }
-                        }
-
-                        if ( GUILayout.Button( "-", GUILayout.Width( 15 ) ) ) {
-                            foreach ( var item in targetlist ) {
-                                Resize( item.objects, Mathf.Max( i + 1, item.objects.Count ), ( ) => new PackagePrefsElement( ) );
-                                item.objects.RemoveAt( i );
-                                objects_count = MinMax.Create( targetlist, v => v.objects.Count );
-                                EditorUtility.SetDirty( item );
-                            }
-                            i--;
-                        }
-                    }
-                }
-                if ( GUILayout.Button( "+", GUILayout.Width( 60 ) ) ) {
-                    foreach ( var item in targetlist ) {
-                        Resize( item.objects, objects_count.max + 1 );
-                        EditorUtility.SetDirty( item );
-                    }
-                }
-                // ↑ Objects
-
-                // ↓ Dynamic Path
-                EditorGUILayout.Separator( );
-                EditorGUILayout.LabelField( t_DynamicPath, EditorStyles.boldLabel );
-                var dpath_count = MinMax.Create( targetlist, v => v.dynamicpath.Count );
-                for ( int i = 0; i < dpath_count.max; i++ ) {
-                    using ( var horizontalScope = new EditorGUILayout.HorizontalScope( ) ) {
-                        // 全てのオブジェクトの値が同じか
-                        bool same = i < dpath_count.min && targetlist.All( v => t.dynamicpath[i] == v.dynamicpath[i] );
-
-                        if ( same ) {
-                            EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-                        } else {
-                            EditorGUILayout.LabelField( new GUIContent( t_Diff_Label, t_Diff_Tooltip ), GUILayout.Width( 30 ) );
-                        }
-
-                        EditorGUI.BeginChangeCheck( );
-                        string path;
-                        if ( same ) {
-                            path = EditorGUILayout.TextField( t.dynamicpath[i] );
-                            path = BrowseButtons( path );
-                        } else {
-                            path = EditorGUILayout.TextField( string.Empty );
-                            path = BrowseButtons( Application.dataPath );
-                        }
-                        EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            foreach ( var item in targetlist ) {
-                                Resize( item.dynamicpath, Mathf.Max( i + 1, item.dynamicpath.Count ) );
-                                item.dynamicpath[i] = path;
-                                dpath_count = MinMax.Create( targetlist, v => v.dynamicpath.Count );
-                                EditorUtility.SetDirty( item );
-                            }
-                        }
-
-                        if ( GUILayout.Button( "-", GUILayout.Width( 15 ) ) ) {
-                            foreach ( var item in targetlist ) {
-                                Resize( item.dynamicpath, Mathf.Max( i + 1, item.dynamicpath.Count ) );
-                                item.dynamicpath.RemoveAt( i );
-                                dpath_count = MinMax.Create( targetlist, v => v.dynamicpath.Count );
-                                EditorUtility.SetDirty( item );
-                            }
-                            i--;
-                        }
-                    }
-                }
-                if ( GUILayout.Button( "+", GUILayout.Width( 60 ) ) ) {
-                    foreach ( var item in targetlist ) {
-                        Resize( item.dynamicpath, dpath_count.max + 1, ( ) => string.Empty );
-                        EditorUtility.SetDirty( item );
-                    }
-                }
-                // ↑ Dynamic Path
-
-                // ↓ Version File
-                EditorGUILayout.Separator();
-                EditorGUILayout.LabelField(t_VersionFile, EditorStyles.boldLabel);
-                using (var horizontalScope = new EditorGUILayout.HorizontalScope())
-                {
-                    var same = targetlist.All(v => t.versionFile.Object == v.versionFile.Object);
-
-                    EditorGUI.BeginChangeCheck();
-                    Object obj;
-                    if (same)
-                    {
-                        obj = EditorGUILayout.ObjectField(t.versionFile.Object, typeof(Object), false);
-                    }
-                    else
-                    {
-                        obj = EditorGUILayout.ObjectField(null, typeof(Object), false);
-                    }
-                    if (EditorGUI.EndChangeCheck())
-                    {
-                        foreach (var item in targetlist)
-                        {
-                            item.versionFile.Object = obj;
-                            EditorUtility.SetDirty(item);
-                        }
-                    }
-
-                    EditorGUI.BeginChangeCheck();
-                    string path;
-                    if (same)
-                    {
-                        path = EditorGUILayout.TextField(t.versionFile.Path);
-                    }
-                    else
-                    {
-                        path = EditorGUILayout.TextField(string.Empty);
-                    }
-                    if (EditorGUI.EndChangeCheck())
-                    {
-                        // パスが変更されたらオブジェクトを置き換える
-                        Object o = AssetDatabase.LoadAssetAtPath<Object>(path);
-                        if (o != null)
-                        {
-                            foreach (var item in targetlist)
-                            {
-                                item.versionFile.Object = o;
-                                EditorUtility.SetDirty(item);
-                            }
-                        }
-                    }
-                }
-                // ↑ Version File
-
-                EditorGUILayout.EndScrollView( );
-
-                // Export Button
-                EditorGUILayout.Separator( );
-                if ( GUILayout.Button( t_Button_ExportPackages ) ) {
-                    UnityPackageExporterEditor.HelpBoxText = string.Empty;
-                    foreach ( var item in targetlist ) {
-                        item.Export( );
-                    }
-                }
-
-                // 出力先一覧
-                EditorGUILayout.Separator( );
-                foreach ( var item in targets ) {
-                    var path = ( item as MizoresPackageExporter ).ExportPath;
-                    EditorGUILayout.LabelField( new GUIContent( path, path ) );
-                }
-                if ( GUILayout.Button( TEXT_BUTTON_OPEN, GUILayout.Width( 60 ) ) ) {
-                    EditorUtility.RevealInFinder( EXPORT_FOLDER_PATH );
-                }
-            }
-            void EditSingle( ) {
-                using ( var s = new EditorGUI.DisabledGroupScope( true ) ) {
-                    EditorGUILayout.ObjectField( t, typeof( MizoresPackageExporter ), false );
-                }
-                scroll = EditorGUILayout.BeginScrollView( scroll );
-                Undo.RecordObject( t, t_Undo );
-
-                // ↓ Objects
-                EditorGUILayout.LabelField( t_Objects, EditorStyles.boldLabel );
-                for ( int i = 0; i < t.objects.Count; i++ ) {
-                    using ( var horizontalScope = new EditorGUILayout.HorizontalScope( ) ) {
-                        PackagePrefsElement item = t.objects[i];
-                        EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-
-                        EditorGUI.BeginChangeCheck( );
-                        item.Object = EditorGUILayout.ObjectField( item.Object, typeof( Object ), false );
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            EditorUtility.SetDirty( t );
-                        }
-
-                        EditorGUI.BeginChangeCheck( );
-                        string path = item.Path;
-                        path = EditorGUILayout.TextField( path );
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            // パスが変更されたらオブジェクトを置き換える
-                            Object o = AssetDatabase.LoadAssetAtPath<Object>( path );
-                            if ( o != null ) {
-                                item.Object = o;
-                            }
-                            EditorUtility.SetDirty( t );
-                        }
-
-                        if ( GUILayout.Button( "-", GUILayout.Width( 15 ) ) ) {
-                            t.objects.RemoveAt( i );
-                            i--;
-                            EditorUtility.SetDirty( t );
-                        }
-                    }
-                }
-                if ( GUILayout.Button( "+", GUILayout.Width( 60 ) ) ) {
-                    t.objects.Add( new PackagePrefsElement( ) );
-                    EditorUtility.SetDirty( t );
-                }
-                // ↑ Objects
-
-                // ↓ Dynamic Path
-                EditorGUILayout.Separator( );
-                EditorGUILayout.LabelField( t_DynamicPath, EditorStyles.boldLabel );
-                for ( int i = 0; i < t.dynamicpath.Count; i++ ) {
-                    using ( var horizontalScope = new EditorGUILayout.HorizontalScope( ) ) {
-                        EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-
-                        EditorGUI.BeginChangeCheck( );
-                        t.dynamicpath[i] = EditorGUILayout.TextField( t.dynamicpath[i] );
-                        t.dynamicpath[i] = BrowseButtons( t.dynamicpath[i] );
-                        EditorGUILayout.LabelField( string.Empty, GUILayout.Width( 30 ) );
-                        if ( EditorGUI.EndChangeCheck( ) ) {
-                            EditorUtility.SetDirty( t );
-                        }
-
-                        if ( GUILayout.Button( "-", GUILayout.Width( 15 ) ) ) {
-                            t.dynamicpath.RemoveAt( i );
-                            i--;
-                            EditorUtility.SetDirty( t );
-                        }
-                    }
-                }
-                if ( GUILayout.Button( "+", GUILayout.Width( 60 ) ) ) {
-                    t.dynamicpath.Add( string.Empty );
-                    EditorUtility.SetDirty( t );
-                }
-                // ↑ Dynamic Path
-
-                // ↓ Version File
-                EditorGUILayout.Separator();
-                EditorGUILayout.LabelField(t_VersionFile, EditorStyles.boldLabel);
-                using (var horizontalScope = new EditorGUILayout.HorizontalScope())
-                {
-                    EditorGUI.BeginChangeCheck();
-                    t.versionFile.Object = EditorGUILayout.ObjectField(t.versionFile.Object, typeof(Object), false);
-                    if (EditorGUI.EndChangeCheck())
-                    {
-                        EditorUtility.SetDirty(t);
-                    }
-                    EditorGUI.BeginChangeCheck();
-                    string path = t.versionFile.Path;
-                    path = EditorGUILayout.TextField(path);
-                    if (EditorGUI.EndChangeCheck())
-                    {
-                        // パスが変更されたらオブジェクトを置き換える
-                        Object o = AssetDatabase.LoadAssetAtPath<Object>(path);
-                        if (o != null)
-                        {
-                            t.versionFile.Object = o;
-                        }
-                        EditorUtility.SetDirty(t);
-                    }
-                }
-                // ↑ Version File
-
-                // Export Button
-                EditorGUILayout.EndScrollView( );
-                EditorGUILayout.Separator( );
-                if ( GUILayout.Button( t_Button_ExportPackage ) ) {
-                    UnityPackageExporterEditor.HelpBoxText = string.Empty;
-                    t.Export( );
-                }
-
-                // Open Button
-                using ( var horizontalScope = new EditorGUILayout.HorizontalScope( ) ) {
-                    EditorGUILayout.LabelField( new GUIContent( t.ExportPath, t.ExportPath ) );
-                    if ( GUILayout.Button( TEXT_BUTTON_OPEN, GUILayout.Width( 60 ) ) ) {
-                        if ( File.Exists( t.ExportPath ) ) {
-                            EditorUtility.RevealInFinder( t.ExportPath );
-                        } else {
-                            EditorUtility.RevealInFinder( EXPORT_FOLDER_PATH );
-                        }
-                    }
-                }
-            }
-            public override void OnInspectorGUI( ) {
-                if ( targets.Length != 1 ) {
-                    EditMultiple( );
-                } else {
-                    EditSingle( );
-                }
-
-                if ( string.IsNullOrEmpty( HelpBoxText ) == false ) {
-                    EditorGUILayout.HelpBox( HelpBoxText.Trim( ), HelpBoxMessageType );
-                }
+        public void OnAfterDeserialize( ) {
+            if ( s_variables != null ) {
+                variables = s_variables.ToDictionary( v => v.key, v => v.value );
             }
         }
-#endif
+
     }
 }
