+++ Assets/Mizore_Nekoyanagi/Util/MultipleEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa44985f139336f448477d0d57280228
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
