+++ Assets/Mizore_Nekoyanagi/Util/PackagePrefsElement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cdf645bf6649b9d4ba97bbbf7b01bbb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
