+++ Assets/Mizore_Nekoyanagi/Util/SingleEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 871a2eb57ce11464583c98b0ff25fb53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
